create IconButtonWithTooltip component
